backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
